Fix multiple embedded resources (see blongden/hal#14).
When a HAL response contains multiple embedded resources of the same
'rel', they are parsed at once.

This fails, as it is an array with separate valid HAL resources in it.
This is fixed by checking if the embeds are an array, and if so, adding
the resources one-by-one.

This is related to blongden/hal#14.

Signed-off-by: Jacob Kiers <jacob@alphacomm.nl>